+++ test/test_utils.py
@@ -128,7 +128,7 @@ def test_construct_filter_validate(osm_return_data: pd.DataFrame, category: Path
     assert set(osm_return_data['@osmId']) == validation_objects[category]
 
 
-def test_fetch_osm_data(expected_compute_input, responses_mock):
+def test_fetch_osm_data(expected_compute_input, default_aoi, responses_mock):
     with open('resources/test/ohsome_line_response.geojson', 'rb') as vector:
         responses_mock.post(
             'https://api.ohsome.org/v1/elements/geometry',
@@ -142,11 +142,11 @@ def test_fetch_osm_data(expected_compute_input, responses_mock):
         geometry=[shapely.LineString([(12.3, 48.22), (12.3, 48.2205), (12.3005, 48.22)])],
         crs=4326,
     )
-    computed_osm_data = fetch_osm_data(expected_compute_input.get_aoi_geom(), 'dummy=yes', OhsomeClient())
+    computed_osm_data = fetch_osm_data(default_aoi, 'dummy=yes', OhsomeClient())
     geopandas.testing.assert_geodataframe_equal(computed_osm_data, expected_osm_data, check_like=True)
 
 
-def test_boost_route_members(expected_compute_input, responses_mock):
+def test_boost_route_members(expected_compute_input, default_aoi, responses_mock):
     with open('resources/test/ohsome_line_response.geojson', 'rb') as vector:
         responses_mock.post(
             'https://api.ohsome.org/v1/elements/geometry',
@@ -194,11 +194,11 @@ def test_boost_route_members(expected_compute_input, responses_mock):
         ],
         crs=4326,
     )
-    computed_output = boost_route_members(expected_compute_input.get_aoi_geom(), paths_input, OhsomeClient())
+    computed_output = boost_route_members(default_aoi, paths_input, OhsomeClient())
     pd.testing.assert_series_equal(computed_output, expected_output)
 
 
-def test_boost_route_members_overlapping_routes(expected_compute_input, responses_mock):
+def test_boost_route_members_overlapping_routes(expected_compute_input, default_aoi, responses_mock):
     with open('resources/test/ohsome_route_response.geojson', 'rb') as vector:
         responses_mock.post(
             'https://api.ohsome.org/v1/elements/geometry',
@@ -214,7 +214,7 @@ def test_boost_route_members_overlapping_routes(expected_compute_input, response
         ],
         crs=4326,
     )
-    computed_output = boost_route_members(expected_compute_input.get_aoi_geom(), paths_input, OhsomeClient())
+    computed_output = boost_route_members(default_aoi, paths_input, OhsomeClient())
     pd.testing.assert_series_equal(computed_output, expected_output)